fix: update code to current version
