+++ project/backend/application/src/team_message/team_message.controller.ts
@@ -2,7 +2,7 @@ import type { Response } from 'express';
 import type { AuthRequest } from '../middleware/auth.middleware.js';
 import { AppError } from '../utils/AppError.js';
 import { emitWithRetries } from '../utils/WebSocketUtils.js';
-import { getIO } from '../server.js';
+import { getIO } from '../utils/socket.js';
 import { getTeamSentMessages, getTeamReceivedMessages, getAllMessages, sendTeamMessage, deleteTeamMessage } from './team_message.service.js';
 
 export const getTeamSentMessagesController = async (req: AuthRequest, res: Response) => {
